pyproject.toml
[tool.poetry]
name = "wsm"
version = "0.1.0"
description = ""
authors = ["Cole Wilson <cole@colewilson.xyz>"]

[tool.poetry.dependencies]
python = "^3.8"
xlsxwriter = "^1.3.7"
pil-compat = "^1.0.0"
eel = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
